Adding energy controllers for cuk

+++ python/lrssoc/cuk/cuk.py
@@ -532,7 +532,7 @@ class Cuk:
             return (-1, status)
 
         t = 1 / freq * np.arange( len(trace_data[0]) )
-        trace_data = np.array(trace_data).T
+        trace_data = np.array(trace_data, dtype=np.float32).T
         
         return (0, (traces, trace_data, t))
